IssueNo:#I5MF2H Description:QuickFixDeploy add log and callback Sig:SIG_ApplicaitonFramework Feature or Bugfix:Feature Binary Source:No Signed-off-by: wangtiantian <wangtiantian19@huawei.com>

+++ services/bundlemgr/src/quick_fix/quick_fix_deployer.cpp
@@ -27,12 +27,20 @@
 namespace OHOS {
 namespace AppExecFwk {
 QuickFixDeployer::QuickFixDeployer(const std::vector<std::string> &bundleFilePaths,
-    std::shared_ptr<QuickFixDataMgr> &quickFixDataMgr) : patchPaths_(bundleFilePaths), quickFixDataMgr_(quickFixDataMgr)
+    const std::shared_ptr<QuickFixDataMgr> &quickFixDataMgr,
+    const sptr<IQuickFixStatusCallback> &statusCallback) : patchPaths_(bundleFilePaths),
+    quickFixDataMgr_(quickFixDataMgr), statusCallback_(statusCallback)
 {}
 
 ErrCode QuickFixDeployer::Execute()
 {
+    if (statusCallback_ == nullptr) {
+        APP_LOGE("QuickFixDeployer statusCallback_ is nullptr");
+        return ERR_APPEXECFWK_QUICK_FIX_PARAM_ERROR;
+    }
     ErrCode ret = DeployQuickFix();
+    deployQuickFixResult_.resultCode = ret;
+    statusCallback_->OnPatchDeployed(deployQuickFixResult_);
     if (ret != ERR_OK) {
         APP_LOGE("QuickFixDeployer errcode %{public}d", ret);
     }
@@ -59,17 +67,17 @@ ErrCode QuickFixDeployer::DeployQuickFix()
             }
         }
     });
-    // 1. parse check multi hqf files, update status DEPLOY_START
+    // parse check multi hqf files, update status DEPLOY_START
     InnerAppQuickFix newInnerAppQuickFix;
     InnerAppQuickFix oldInnerAppQuickFix;
     ret = ToDeployStartStatus(realFilePaths, newInnerAppQuickFix, oldInnerAppQuickFix);
     if (ret != ERR_OK) {
         return ret;
     }
-    // 2. extract diff files, apply diff patch and copy hqf, update status DEPLOY_END
+    // extract diff files, apply diff patch and copy hqf, update status DEPLOY_END
     ret = ToDeployEndStatus(newInnerAppQuickFix, oldInnerAppQuickFix);
     if (ret != ERR_OK) {
-        bool isExist = oldInnerAppQuickFix.GetAppQuickFix().bundleName.empty();
+        bool isExist = !oldInnerAppQuickFix.GetAppQuickFix().bundleName.empty();
         if (isExist) {
             quickFixDataMgr_->SaveInnerAppQuickFix(oldInnerAppQuickFix);
         } else {
@@ -77,8 +85,7 @@ ErrCode QuickFixDeployer::DeployQuickFix()
         }
         return ret;
     }
-    // 3. callback
-    // 4. remove old deploying patch_versionCode
+    // remove old deploying patch_versionCode
     const AppQuickFix &appQuick = oldInnerAppQuickFix.GetAppQuickFix();
     if (!appQuick.bundleName.empty()) {
         std::string oldPath = Constants::BUNDLE_CODE_DIR + Constants::PATH_SEPARATOR +
@@ -100,60 +107,76 @@ ErrCode QuickFixDeployer::ToDeployStartStatus(const std::vector<std::string> &bu
 {
     APP_LOGD("ToDeployStartStatus start.");
     if (quickFixDataMgr_ == nullptr) {
+        APP_LOGE("error: quickFixDataMgr_ is nullptr");
         return ERR_APPEXECFWK_QUICK_FIX_INTERNAL_ERROR;
     }
     std::unordered_map<std::string, AppQuickFix> infos;
-    // 1.parse and check multi app quick fix info
+    // parse and check multi app quick fix info
     ErrCode ret = ParseAndCheckAppQuickFixInfos(bundleFilePaths, infos);
     if (ret != ERR_OK) {
         return ret;
     }
     const AppQuickFix &appQuickFix = infos.begin()->second;
+    ToDeployQuickFixResult(appQuickFix);
     bool isExist = quickFixDataMgr_->QueryInnerAppQuickFix(appQuickFix.bundleName, oldInnerAppQuickFix);
     const AppQuickFix &oldAppQuickFix = oldInnerAppQuickFix.GetAppQuickFix();
     // exist and type same need to check version code
     if (isExist && (appQuickFix.deployingAppqfInfo.type == oldAppQuickFix.deployingAppqfInfo.type)) {
-        // 2. check current app quick fix version code
+        // check current app quick fix version code
         ret = CheckPatchVersionCode(appQuickFix, oldAppQuickFix);
         if (ret != ERR_OK) {
             return ret;
         }
     }
-    // 3. check bundleName exist
+    // check bundleName exist
     BundleInfo bundleInfo;
     ret = GetBundleInfo(appQuickFix.bundleName, bundleInfo);
     if (ret != ERR_OK) {
+        APP_LOGE("GetBundleInfo failed, bundleName: %{public}s", appQuickFix.bundleName.c_str());
         return ret;
     }
-    // 4. check with installed bundle
+    // check with installed bundle
     if (appQuickFix.deployingAppqfInfo.type == QuickFixType::PATCH) {
         ret = ProcessPatchDeployStart(bundleFilePaths, bundleInfo, infos);
-        if (ret != ERR_OK) {
-            APP_LOGE("ProcessPatchDeployStart failed, errcode:%{public}d", ret);
-            return ret;
-        }
     } else if (appQuickFix.deployingAppqfInfo.type == QuickFixType::HOT_RELOAD) {
-        ret = ProcessHotReloadDeployStart(bundleInfo, infos);
-        if (ret != ERR_OK) {
-            APP_LOGE("ProcessPatchDeployStart failed, errcode:%{public}d", ret);
-            return ret;
-        }
+        ret = ProcessHotReloadDeployStart(bundleInfo, appQuickFix);
     } else {
         APP_LOGE("unknown quick fix type");
         return ERR_APPEXECFWK_QUICK_FIX_UNKNOWN_QUICK_FIX_TYPE;
     }
-    // 5. convert to InnerAppQuickFix
-    ret = ToInnerAppQuickFix(infos, oldInnerAppQuickFix, newInnerAppQuickFix);
     if (ret != ERR_OK) {
+        APP_LOGE("Process Patch or HotReload DeployStart failed, errcode:%{public}d", ret);
         return ret;
     }
-    // 6. save infos and update status DEPLOY_START
+    // convert to InnerAppQuickFix
+    ret = ToInnerAppQuickFix(infos, oldInnerAppQuickFix, newInnerAppQuickFix);
+    if (ret != ERR_OK) {
+        APP_LOGE("ToInnerAppQuickFix failed, errcode: %{public}d", ret);
+        return ret;
+    }
+    // save infos and update status DEPLOY_START
     ret = SaveAppQuickFix(newInnerAppQuickFix);
     if (ret != ERR_OK) {
+        APP_LOGE("SaveAppQuickFix failed, errcode: %{public}d", ret);
         return ret;
     }
     APP_LOGD("ToDeployStartStatus end.");
     return ERR_OK;
+}
+
+void QuickFixDeployer::ToDeployQuickFixResult(const AppQuickFix &appQuickFix)
+{
+    APP_LOGD("ToDeployQuickFixResult start.");
+    deployQuickFixResult_.bundleName = appQuickFix.bundleName;
+    deployQuickFixResult_.bundleVersionCode = appQuickFix.versionCode;
+    deployQuickFixResult_.patchVersionCode = appQuickFix.deployingAppqfInfo.versionCode;
+    deployQuickFixResult_.type = appQuickFix.deployingAppqfInfo.type;
+    deployQuickFixResult_.isSoContained = !appQuickFix.deployingAppqfInfo.nativeLibraryPath.empty();
+    deployQuickFixResult_.moduleNames.clear();
+    for (const auto &hqf : appQuickFix.deployingAppqfInfo.hqfInfos) {
+        deployQuickFixResult_.moduleNames.emplace_back(hqf.moduleName);
+    }
+    APP_LOGD("ToDeployQuickFixResult end.");
 }
 
 ErrCode QuickFixDeployer::ProcessPatchDeployStart(
@@ -162,24 +185,33 @@ ErrCode QuickFixDeployer::ProcessPatchDeployStart(
     std::unordered_map<std::string, AppQuickFix> &infos)
 {
     if (infos.empty()) {
+        APP_LOGE("error: appQuickFix infos is empty");
         return ERR_APPEXECFWK_QUICK_FIX_PROFILE_PARSE_FAILED;
     }
     QuickFixChecker checker;
-    // 1. check multiple cpuAbi and native library path
+    // check multiple cpuAbi and native library path
     ErrCode ret = checker.CheckMultiNativeSo(infos);
     if (ret != ERR_OK) {
+        APP_LOGE("ProcessPatchDeployStart check native so failed");
         return ret;
     }
     // parse signature info
     std::vector<Security::Verify::HapVerifyResult> hapVerifyRes;
-    ret = checker.CheckMultipleHapsSignInfo(bundleFilePaths, hapVerifyRes);
+    ret = checker.CheckMultipleHqfsSignInfo(bundleFilePaths, hapVerifyRes);
     if (ret != ERR_OK) {
+        APP_LOGE("ProcessPatchDeployStart check check multiple hqfs signInfo failed");
         return ret;
     }
+    if (hapVerifyRes.empty()) {
+        APP_LOGE("error: appQuickFix hapVerifyRes is empty");
+        return ERR_APPEXECFWK_INSTALL_FAILED_INCOMPATIBLE_SIGNATURE;
+    }
+    const auto &provisionInfo = hapVerifyRes[0].GetProvisionInfo();
     const AppQuickFix &appQuickFix = infos.begin()->second;
-    // 4. check with installed bundle
-    ret = CheckWithInstalledBundle(appQuickFix, bundleInfo, hapVerifyRes);
+    // check with installed bundle, signature info, bundleName, versionCode
+    ret = checker.CheckPatchWithInstalledBundle(appQuickFix, bundleInfo, provisionInfo);
     if (ret != ERR_OK) {
+        APP_LOGE("check AppQuickFixInfos with installed bundle failed, errcode : %{public}d", ret);
         return ret;
     }
     return ERR_OK;
@@ -187,13 +219,9 @@ ErrCode QuickFixDeployer::ProcessPatchDeployStart(
 
 ErrCode QuickFixDeployer::ProcessHotReloadDeployStart(
     const BundleInfo &bundleInfo,
-    const std::unordered_map<std::string, AppQuickFix> &infos)
+    const AppQuickFix &appQuickFix)
 {
-    if (infos.empty()) {
-        return ERR_APPEXECFWK_QUICK_FIX_PROFILE_PARSE_FAILED;
-    }
     QuickFixChecker checker;
-    const AppQuickFix &appQuickFix = infos.begin()->second;
     ErrCode ret = checker.CheckHotReloadWithInstalledBundle(appQuickFix, bundleInfo);
     if (ret != ERR_OK) {
         APP_LOGE("check AppQuickFixInfos with installed bundle failed");
@@ -222,15 +250,18 @@ ErrCode QuickFixDeployer::ToDeployEndStatus(InnerAppQuickFix &newInnerAppQuickFi
     } else if (newQuickFix.deployingAppqfInfo.type == QuickFixType::HOT_RELOAD) {
         ret = ProcessHotReloadDeployEnd(newQuickFix, newPatchPath);
     } else {
+        APP_LOGE("error: unknown QuickFixType");
         return ERR_APPEXECFWK_QUICK_FIX_PROFILE_PARSE_FAILED;
     }
     if (ret != ERR_OK) {
-        APP_LOGE("ToDeployEndStatus failed, bundleName:%{public}s", newQuickFix.bundleName.c_str());
+        APP_LOGE("Process Patch or HotReload DeployEnd failed, bundleName:%{public}s",
+            newQuickFix.bundleName.c_str());
         return ret;
     }
     // move hqf files to new patch path
     ret = MoveHqfFiles(newInnerAppQuickFix, newPatchPath);
     if (ret != ERR_OK) {
+        APP_LOGE("error MoveHqfFiles failed, bundleName: %{public}s", newQuickFix.bundleName.c_str());
         return ret;
     }
     // save and update status DEPLOY_END
@@ -253,7 +284,19 @@ ErrCode QuickFixDeployer::ProcessPatchDeployEnd(const AppQuickFix &appQuickFix, 
         return ret;
     }
     if (!appQuickFix.deployingAppqfInfo.nativeLibraryPath.empty()) {
-        // 1. extract diff so, diff so path
+        const std::string &libraryPath = appQuickFix.deployingAppqfInfo.nativeLibraryPath;
+        std::string oldSoPath = Constants::BUNDLE_CODE_DIR + Constants::PATH_SEPARATOR + appQuickFix.bundleName +
+            Constants::PATH_SEPARATOR + libraryPath;
+        APP_LOGD("ProcessPatchDeployEnd oldPath %{public}s", oldSoPath.c_str());
+        bool pathExist = false;
+        // if old so path does not exist then return ERR_OK
+        ret = InstalldClient::GetInstance()->IsExistDir(oldSoPath, pathExist);
+        if (!pathExist && (ret == ERR_OK)) {
+            APP_LOGD("bundleName: %{public}s no so path", appQuickFix.bundleName.c_str());
+            deployQuickFixResult_.isSoContained = false;
+            return ERR_OK;
+        }
+        // extract diff so, diff so path
         std::string diffFilePath = Constants::HAP_COPY_PATH + Constants::PATH_SEPARATOR +
             appQuickFix.bundleName + Constants::TMP_SUFFIX;
         ScopeGuard guardRemoveDiffPath([diffFilePath] { InstalldClient::GetInstance()->RemoveDir(diffFilePath); });
@@ -261,12 +304,17 @@ ErrCode QuickFixDeployer::ProcessPatchDeployEnd(const AppQuickFix &appQuickFix, 
         if (ret != ERR_OK) {
             return ret;
         }
-        // 2. apply diff patch
-        ret = ApplyDiffPatch(appQuickFix.bundleName, appQuickFix.deployingAppqfInfo.nativeLibraryPath,
-            diffFilePath, patchPath);
+        // apply diff patch
+        std::string newSoPath = patchPath + Constants::PATH_SEPARATOR + libraryPath;
+        ret = InstalldClient::GetInstance()->ApplyDiffPatch(oldSoPath, diffFilePath, newSoPath);
         if (ret != ERR_OK) {
             APP_LOGE("ApplyDiffPatch failed, bundleName:%{public}s", appQuickFix.bundleName.c_str());
             return ret;
+        }
+        ret = InstalldClient::GetInstance()->IsExistDir(newSoPath, deployQuickFixResult_.isSoContained);
+        if (ret != ERR_OK) {
+            APP_LOGW("InstalldClient::IsExistDir failed, errcode: %{public}d", ret);
+            deployQuickFixResult_.isSoContained = true;
         }
     }
     return ERR_OK;
@@ -322,38 +370,19 @@ ErrCode QuickFixDeployer::GetBundleInfo(const std::string &bundleName, BundleInf
 {
     std::shared_ptr<BundleMgrService> bms = DelayedSingleton<BundleMgrService>::GetInstance();
     if (bms == nullptr) {
+        APP_LOGE("error: bms is nullptr");
         return ERR_APPEXECFWK_QUICK_FIX_INTERNAL_ERROR;
     }
     std::shared_ptr<BundleDataMgr> dataMgr = bms->GetDataMgr();
     if (dataMgr == nullptr) {
+        APP_LOGE("error: dataMgr is nullptr");
         return ERR_APPEXECFWK_QUICK_FIX_INTERNAL_ERROR;
     }
     // check bundleName is exists
     if (!dataMgr->GetBundleInfo(bundleName, BundleFlag::GET_BUNDLE_WITH_APPQF_INFO,
         bundleInfo, Constants::ANY_USERID)) {
+        APP_LOGE("error: GetBundleInfo failed, bundleName: %{public}s not exist", bundleName.c_str());
         return ERR_APPEXECFWK_QUICK_FIX_BUNDLE_NAME_NOT_EXIST;
-    }
-    return ERR_OK;
-}
-
-ErrCode QuickFixDeployer::CheckWithInstalledBundle(
-    const AppQuickFix &appQuickFix,
-    const BundleInfo &bundleInfo,
-    const std::vector<Security::Verify::HapVerifyResult> &hapVerifyRes)
-{
-    if (hapVerifyRes.empty()) {
-        return ERR_APPEXECFWK_INSTALL_FAILED_INCOMPATIBLE_SIGNATURE;
-    }
-    QuickFixChecker checker;
-    // hot reload does not require signature verify
-    ErrCode ret = checker.CheckSignatureInfo(bundleInfo, hapVerifyRes[0].GetProvisionInfo());
-    if (ret != ERR_OK) {
-        return ret;
-    }
-    ret = checker.CheckPatchWithInstalledBundle(appQuickFix, bundleInfo);
-    if (ret != ERR_OK) {
-        APP_LOGE("check AppQuickFixInfos with installed bundle failed");
-        return ret;
     }
     return ERR_OK;
 }
@@ -362,6 +391,7 @@ ErrCode QuickFixDeployer::ToInnerAppQuickFix(const std::unordered_map<std::strin
     const InnerAppQuickFix &oldInnerAppQuickFix, InnerAppQuickFix &newInnerAppQuickFix)
 {
     if (infos.empty()) {
+        APP_LOGE("error: appQuickFix is empty");
         return ERR_APPEXECFWK_QUICK_FIX_INTERNAL_ERROR;
     }
     AppQuickFix oldAppQuickFix = oldInnerAppQuickFix.GetAppQuickFix();
@@ -374,10 +404,13 @@ ErrCode QuickFixDeployer::ToInnerAppQuickFix(const std::unordered_map<std::strin
     mark.status = QuickFixStatus::DEPLOY_START;
     for (auto iter = infos.begin(); iter != infos.end(); ++iter) {
         const auto &quickFix = iter->second;
+        // hqfInfos will not be empty, it has been judged before.
+        const std::string &moduleName = quickFix.deployingAppqfInfo.hqfInfos[0].moduleName;
         if (!newInnerAppQuickFix.AddHqfInfo(quickFix)) {
+            APP_LOGE("error: appQuickFix add hqf moduleName: %{public}s failed", moduleName.c_str());
             return ERR_APPEXECFWK_QUICK_FIX_ADD_HQF_FAILED;
         }
-        mark.moduleName = quickFix.deployingAppqfInfo.hqfInfos[0].moduleName;
+        mark.moduleName = moduleName;
     }
     newInnerAppQuickFix.SetQuickFixMark(mark);
     return ERR_OK;
@@ -400,15 +433,15 @@ ErrCode QuickFixDeployer::CheckPatchVersionCode(
 
 ErrCode QuickFixDeployer::SaveAppQuickFix(const InnerAppQuickFix &innerAppQuickFix)
 {
-    APP_LOGD("SaveAppQuickFix start.");
     if (quickFixDataMgr_ == nullptr) {
+        APP_LOGE("error: quickFixDataMgr_ is nullptr");
         return ERR_APPEXECFWK_QUICK_FIX_INTERNAL_ERROR;
     }
     if (!quickFixDataMgr_->SaveInnerAppQuickFix(innerAppQuickFix)) {
-        APP_LOGE("inner app quick fix save failed");
+        APP_LOGE("bundleName: %{public}s, inner app quick fix save failed",
+            innerAppQuickFix.GetAppQuickFix().bundleName.c_str());
         return ERR_APPEXECFWK_QUICK_FIX_SAVE_APP_QUICK_FIX_FAILED;
     }
-    APP_LOGD("SaveAppQuickFix end.");
     return ERR_OK;
 }
 
@@ -421,13 +454,13 @@ ErrCode QuickFixDeployer::ExtractDiffFiles(const std::string &targetPath,
     }
     for (const auto &hqf : appQfInfo.hqfInfos) {
         if (hqf.hapFilePath.empty()) {
-            APP_LOGE("error hapFilePath is empty");
+            APP_LOGE("error: hapFilePath is empty");
             return ERR_APPEXECFWK_QUICK_FIX_PARAM_ERROR;
         }
         // extract so to targetPath
         ErrCode ret = InstalldClient::GetInstance()->ExtractDiffFiles(hqf.hapFilePath, targetPath, appQfInfo.cpuAbi);
         if (ret != ERR_OK) {
-            APP_LOGE("ExtractDiffFiles failed");
+            APP_LOGE("error: ExtractDiffFiles failed");
             return ret;
         }
     }
@@ -435,32 +468,11 @@ ErrCode QuickFixDeployer::ExtractDiffFiles(const std::string &targetPath,
     return ERR_OK;
 }
 
-ErrCode QuickFixDeployer::ApplyDiffPatch(
-    const std::string &bundleName,
-    const std::string &libraryPath,
-    const std::string &diffFilePath,
-    const std::string &newSoPath)
-{
-    APP_LOGD("ApplyDiffPatch start.");
-    if (bundleName.empty()|| libraryPath.empty() || diffFilePath.empty() || newSoPath.empty()) {
-        APP_LOGD("ApplyDiffPatch no diff files.");
-        return ERR_OK;
-    }
-    std::string oldSoPath = Constants::BUNDLE_CODE_DIR + Constants::PATH_SEPARATOR + bundleName +
-        Constants::PATH_SEPARATOR + libraryPath;
-    ErrCode ret = InstalldClient::GetInstance()->ApplyDiffPatch(oldSoPath, diffFilePath, newSoPath);
-    if (ret != ERR_OK) {
-        APP_LOGE("ApplyDiffPatch failed");
-        return ret;
-    }
-    APP_LOGD("ApplyDiffPatch end.");
-    return ERR_OK;
-}
-
 ErrCode QuickFixDeployer::MoveHqfFiles(InnerAppQuickFix &innerAppQuickFix, const std::string &targetPath)
 {
     APP_LOGD("MoveHqfFiles start.");
     if (targetPath.empty() || (quickFixDataMgr_ == nullptr)) {
+        APP_LOGE("MoveHqfFiles params error");
         return ERR_APPEXECFWK_QUICK_FIX_PARAM_ERROR;
     }
     QuickFixMark mark = innerAppQuickFix.GetQuickFixMark();
@@ -477,7 +489,7 @@ ErrCode QuickFixDeployer::MoveHqfFiles(InnerAppQuickFix &innerAppQuickFix, const
         std::string realPath = path + info.moduleName + Constants::QUICK_FIX_FILE_SUFFIX;
         ErrCode ret = InstalldClient::GetInstance()->MoveFile(info.hapFilePath, realPath);
         if (ret != ERR_OK) {
-            APP_LOGE("MoveFile failed");
+            APP_LOGE("error MoveFile failed");
             return ret;
         }
         info.hapFilePath = realPath;
